+++ doc/reference/CHANGELOG.md
@@ -23,6 +23,7 @@ Updated components will trigger pod rerolls at upgrade.
 
 - cert-manager: dependency removal complete. If you deployed a Software Factory with sf-operator prior to version
   v0.0.55 please upgrade to [v0.0.55](#v0055---2025-03-03) first to ensure proper removal of dangling cert-manager-related resources.
+- The spec config-location "base-url" attribute is no longer required.
 
 ### Fixed